+++ src/req_wrap.h
@@ -7,6 +7,9 @@
 #include "util.h"
 #include "v8.h"
 
+#include <sys/syscall.h>
+#include <unistd.h>
+
 namespace node {
 
 class Environment;
@@ -52,7 +55,7 @@ class ReqWrap : public AsyncWrap, public ReqWrapBase {
                   double execution_async_id = kInvalidAsyncId,
                   bool silent = false) override {
     AsyncWrap::AsyncReset(resource, execution_async_id, silent);
-    syscall(516, get_async_id(), req_);
+    syscall(516, static_cast<int64_t>(get_async_id()), req_);
   }
 
  private:
